requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
sqlalchemy>=2.0.0
faker>=22.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
